Backporting

+++ src/Basescape/SellState.h
@@ -65,9 +65,9 @@ private:
 	/// Gets the Type of the selected item.
 	enum SellType getType(size_t selected) const;
 	/// Gets the index of selected item.
-	int getItemIndex(size_t selected) const;
+	size_t getItemIndex(size_t selected) const;
 	/// Gets the index of the selected craft.
-	int getCraftIndex(size_t selected) const;
+	size_t getCraftIndex(size_t selected) const;
 public:
 	/// Creates the Sell state.
 	SellState(Game *game, Base *base, OptionsOrigin origin = OPT_GEOSCAPE);
