+++ src/Ruleset/MapDataSet.cpp
@@ -71,7 +71,7 @@ std::string MapDataSet::getName() const
  * Gets the MapDataSet size.
  * @return The size in number of records.
  */
-int MapDataSet::getSize() const
+size_t MapDataSet::getSize() const
 {
 	return _objects.size();
 }
