+++ src/Ruleset/UfoTrajectory.cpp
@@ -70,7 +70,7 @@ UfoTrajectory::UfoTrajectory(const std::string &id) : _id(id), _groundTimer(5)
 void UfoTrajectory::load(const YAML::Node &node)
 {
 	_id = node["id"].as<std::string>(_id);
-	_groundTimer = node["groundTimer"].as<unsigned>(_groundTimer);
+	_groundTimer = node["groundTimer"].as<size_t>(_groundTimer);
 	_waypoints = node["waypoints"].as< std::vector<TrajectoryWaypoint> >(_waypoints);
 }
 
@@ -79,7 +79,7 @@ void UfoTrajectory::load(const YAML::Node &node)
  * @param wp The waypoint.
  * @return The altitude.
  */
-std::string UfoTrajectory::getAltitude(unsigned wp) const
+std::string UfoTrajectory::getAltitude(size_t wp) const
 {
 	return altitudeString[_waypoints[wp].altitude];
 }
